Add files via upload
Electric vehicles (EV) are revolutionizing the automotive industry by offering a Cleaner and 
more sustainable mode of transport but the availability of charging Stations are less and it takes 
8-9 hours charge a 60kWh battery. This study explores the feasibility and efficiency of charging 
electric vehicle (EV) batteries through electromagnetic induction generated by the vehicle's 
motion. By this concept, there will be a permanent primary battery which is charged by normal 
charging method, and along with this there will be a secondary battery which is get charged by 
the concept of reverse engineering.  The reverse engineering is the primary method that we are going to implement in this  project. This mainly focuses on driving charge from the dc motor that is attached to front wheel of the vehicle
When a DC motor is rotated (e.g., by the vehicle’s front wheel), it can act as a generator.
 As the motor spins, it generates electrical energy (a voltage). The generated voltage can be used to charge a battery.
 The battery stores the electrical energy for later use.
The paper discusses the design considerations, efficiency factors, and potential benefits of this innovative 
charging method, including reduced reliance on stationary charging stations and increased 
convenience for EV owners. Experimental results and simulations demonstrate the 
effectiveness of the proposed approach, paving the way for sustainable and efficient charging 
solutions for electric vehicles.